+++ binding.gyp
@@ -3,7 +3,11 @@
     {
       "target_name": "libvirt",
       'product_prefix': 'lib',
-      "sources": [ "generated/libvirt.c" ],
+      "sources": [ 
+        'generated/libvirt.c',
+        'generated/connect.c',
+        'generated/domain.c',
+      ],
       'conditions': [
         ['OS!="win"', {
           'link_settings': {
