+++ src/manual/libvirt.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "connect.h"
+#include "domain.h"
+
