+++ src/override/connect.h
@@ -1,0 +1,1 @@
+napi_value libvirt_virConnectOpen(napi_env env, napi_callback_info info);
