+++ src/override/domain.h
@@ -1,0 +1,2 @@
+napi_value libvirt_virDomainLookupByName(napi_env env, napi_callback_info info);
+napi_value libvirt_virDomainGetID(napi_env env, napi_callback_info info);
